refactor(OptimizeAndTool): 加固全图透视、伤害显示、饰品盒与 QoL 系统边界

+++ Mods/OptimizeAndTool/OptimizeAndTool/OptimizeAndTool/Content/QoL/Economy.cs
@@ -45,7 +45,7 @@ namespace OptimizeAndTool.Content.QoL
             if (!Economy.EnableCoinDropRate.val) return;
             if (This.value > 0f)
             {
-                This.value = (int)(This.value * Economy.CoinDropRate.val);
+                This.value = This.value * Economy.CoinDropRate.val;
             }
         }
     }
